Paranuara/requirements.txt
asgiref==3.2.7
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.6
django-cors-headers==3.3.0
django-mysql==3.5.0
djangorestframework==3.11.0
drf-yasg==1.17.1
idna==2.9
importlib-metadata==1.6.0
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
MarkupSafe==1.1.1
mysql==0.0.2
mysqlclient==1.4.6
nltk==3.5
openapi-codec==1.3.2
packaging==20.4
pathlib==1.0.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.5.14
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
tqdm==4.46.0
uritemplate==3.0.1
urllib3==1.25.9
zipp==3.1.0
